AUT-5466: Emit passkey verification successful result from finish passkey handler
To come in subsequent commit: properly deriving the allowedcredentials list

+++ audit-events/src/main/java/uk/gov/di/authentication/auditevents/entity/shared/passkeys/PasskeyDetail.java
@@ -2,14 +2,14 @@ package uk.gov.di.authentication.auditevents.entity.shared.passkeys;
 
 public record PasskeyDetail(
         PasskeyAuthenticationRequest passkeyAuthenticationRequest,
-        int passkeyCounter,
+        long passkeyCounter,
         boolean passkeyCredentialBackedUp,
         String passkeyCredentialDeviceType,
         boolean passkeyUserVerified,
         String passkeyVerificationFailureReason) {
     public static PasskeyDetail verificationFailed(
             String userVerification,
-            int passkeyCounter,
+            long passkeyCounter,
             boolean passkeyCredentialBackedUp,
             String passkeyCredentialDeviceType,
             String failureReason) {
@@ -25,7 +25,7 @@ public record PasskeyDetail(
 
     public static PasskeyDetail verificationSuccessful(
             String userVerification,
-            int passkeyCounter,
+            long passkeyCounter,
             boolean passkeyCredentialBackedUp,
             String passkeyCredentialDeviceType) {
         var userVerified = true;
@@ -39,7 +39,7 @@ public record PasskeyDetail(
     }
 
     public static PasskeyDetail authenticationSuccessful(
-            int passkeyCounter,
+            long passkeyCounter,
             boolean passkeyCredentialBackedUp,
             String passkeyCredentialDeviceType) {
         var userVerified = false;
